Move 'internal/verify' SemVer functions to 'internal/semver'.
